refactor(rdf-core): the last xsd_regex cursor folds over the scanner, and translate.rs goes
`ecma_262_divergences` was the last hand-written character cursor in the
xsd_regex module. It moves to a new private `ecma` module and becomes a flat
fold over the `Scanner` token stream: no `chars.get(i)`, no index advance, no
hand-rolled `\p{` name scan. Nine match arms, cyclomatic complexity roughly 11
(was ~28). With it gone, `translate.rs` held nothing and is deleted, leaving
the module with exactly one character cursor, in `scan.rs`.

Two behaviour-preservation details:

- An unterminated `\p{Is…` was reported by the old cursor as a block-escape
  divergence, but the scanner rejects it outright and emits no
  `UnicodeProperty` token. The corpus pins that exact input (`^\p{IsBasicLatin`
  in block-escapes.cases), so the fold recovers the divergence from the scan
  failure's `source_text()` via a `\p{Is`/`\P{Is` prefix check. No other
  rejected construct's source begins `\p{`/`\P{`.
- The moved private function's real return type was `Vec<&'static str>`, not
  `Vec<String>`; it was kept as-is so the change is purely cursor removal, and
  the public `ecma_262_divergences` wrapper in `mod.rs` still returns a
  byte-identical `Vec<String>`.

Behaviour is unchanged, verified with a temporary differential test (since
deleted) comparing old and new over 250 inputs: all 197 corpus patterns from
the six .cases files plus 53 hand-written constructs covering `\i \I \c \C`,
`\s \S \w \W`, `\d \D`, `\p{Is…}`/`\P{Is…}`, category escapes, class
subtraction including negated and nested, dot inside and outside classes,
groups, backreferences and malformed forms. The vectors were identical
including order; zero divergence.

Verified: `cargo test -p purrdf-core`; `cargo test -p purrdf-shapes --lib`
(637 passed); `cargo clippy -p purrdf-core -p purrdf-shapes --all-targets`;
`cargo fmt --check`; `RUSTDOCFLAGS="-D warnings" cargo doc -p purrdf-core
--no-deps`; `bash scripts/check-generated.sh`. EXPECTED_CASES is still 197 and
generated/shapes-loss-ledger.json is unchanged, so the divergence output did
not move.
